order-service-implemented

+++ order-service/src/main/java/com/vendor_marketplace/order_service/dao/implementation/OrderDaoImpl.java
@@ -1,34 +1,44 @@
 package com.vendor_marketplace.order_service.dao.implementation;
 
+import com.vendor_marketplace.common.dto.enums.OrderStatus;
+import com.vendor_marketplace.common.dto.enums.PaymentStatus;
 import com.vendor_marketplace.common.exception.ResourceNotFoundException;
 import com.vendor_marketplace.order_service.dao.interfaces.OrderDao;
 import com.vendor_marketplace.order_service.dao.repository.OrderRepository;
 import com.vendor_marketplace.order_service.models.entity.Order;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 class OrderDaoImpl implements OrderDao {
 
     private final OrderRepository orderRepository;
 
     @Override
-    public void save(Order order) {
-        orderRepository.save(order);
+    public Order save(Order order) {
+        return orderRepository.save(order);
+    }
+
+    @Override
+    public List<Order> saveAll(List<Order> orders) {
+       return orderRepository.saveAll(orders);
     }
 
     // for admin  :> in future adding multiple sort filter for admin
     @Override
-    public Page<Order> findAll(int page, int size) {
-        Sort sort = Sort.by(Sort.Direction.DESC, "createdAt");
+    public Page<Order> findAll(int page, int size, boolean isNewest) {
+        Sort sort = Sort.by(isNewest ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt");
         Pageable pageable = PageRequest.of(page, size, sort);
         return orderRepository.findAll(pageable);
     }
@@ -60,12 +70,48 @@ class OrderDaoImpl implements OrderDao {
 
     @Override
     public void deleteById(Long id) {
-        orderRepository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Order not found")
-        );
+        log.info("🗑Deleting orders | order Id={}", id);
+        if (!orderRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Order not found with id: " + id);
+        }
 
         orderRepository.deleteById(id);
+        log.info("Deleted successfully : order id: {}", id);
     }
 
+    @Override
+    public void deleteOrderIdOrders(String orderId) {
+        log.info("🗑Deleting orders | orderId={}", orderId);
+        long orderCount = orderRepository.countByOrderId(orderId);
 
+        if (orderCount == 0) {
+            log.error("No orders found | orderId={}", orderId);
+            throw new ResourceNotFoundException("Order not found with id: " + orderId);
+        }
+        log.info("Orders to delete | orderId={} | count={}", orderId, orderCount);
+        int deletedCount = orderRepository.deleteByOrderId(orderId);
+        log.info("Delete completed | orderId={} | deletedCount={}", orderId, deletedCount);
+
+    }
+
+    @Override
+    public int updateOrderAndPaymentStatus(String orderId, OrderStatus orderStatus, PaymentStatus paymentStatus
+    ) {
+        return orderRepository.bulkUpdateOrders(orderId, orderStatus, paymentStatus, LocalDateTime.now());
+    }
+
+    @Override
+    public boolean existByOrderId(String orderId) {
+        return orderRepository.existsByOrderId(orderId);
+    }
+
+    @Override
+    public boolean existByOrderIdAndCreatedAtBefore(String orderId, LocalDateTime cutoffTime) {
+        return orderRepository.existsByOrderIdAndCreatedAtBefore(orderId, cutoffTime);
+    }
+
+    @Override
+    public List<Order> findByOrderIdAndUserId(String orderId, String userId) {
+        return orderRepository.findByOrderIdAndUserId(orderId, userId);
+    }
 }
